+++ src/babashka/impl/classes.clj
@@ -122,6 +122,15 @@
                {:name "setAccessible"}
                {:name "get"}
                {:name "set"}]}
+    java.lang.reflect.Constructor
+    {:methods [{:name "getName"}
+               {:name "getModifiers"}
+               {:name "getParameterTypes"}
+               {:name "getParameterCount"}]}
+    java.lang.reflect.Executable
+    {:methods [{:name "getParameterCount"}]}
+    java.util.stream.Collectors
+    {:methods [{:name "toList"}]}
     java.lang.reflect.Array
     {:methods [{:name "newInstance"}
                {:name "set"}]}
@@ -748,6 +757,7 @@
                       clojure.lang.IPersistentStack
                       clojure.lang.IPersistentList
                       clojure.lang.IRecord
+                      clojure.lang.IType
                       clojure.lang.IReduce
                       clojure.lang.IReduceInit
                       clojure.lang.IKVReduce
@@ -793,6 +803,7 @@
                       java.lang.ExceptionInInitializerError
                       java.lang.LinkageError
                       java.lang.ThreadDeath
+                      java.lang.StackOverflowError
                       java.lang.VirtualMachineError
                       java.lang.NoSuchFieldException
                       java.sql.Timestamp
@@ -1051,6 +1062,7 @@
     ProcessBuilder java.lang.ProcessBuilder
     SecurityException java.lang.SecurityException
     Short java.lang.Short
+    StackOverflowError java.lang.StackOverflowError
     StackTraceElement java.lang.StackTraceElement
     String java.lang.String
     StringBuilder java.lang.StringBuilder